requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.0.1
pytest==7.4.3
hypothesis==6.88.1
groq>=1.1.0
